chore: organize story files of templates
